Fix mergeSort algorithm.
I was facing problems when the array had more than 4 elements, after
hours of debugging i found out the issue was because I WASN'T ACTUALLY
RETURNING THE VALUE!!!
I was just logging it to the console the whole time. Crazy work. I need
to sharpen my problem solving and debugging skills. Hmm, saying it like
that is a little redundant because debugging is also problem solving.
Welp!